LPS-53183 Change module name and packages
