+++ modules/core/osgi-service-tracker-map/.classpath
