+++ modules/core/osgi-service-tracker-map/.project
