+++ modules/core/osgi-service-tracker-map/build.xml
