+++ modules/core/osgi-service-tracker-map/ivy.xml
