+++ modules/core/osgi-service-tracker-map/test/integration/com/liferay/osgi/service/tracker/map/internal/TrackedTwo.java
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.osgi.service.tracker.map.internal;
+
+/**
+ * @author Carlos Sierra Andrés
+ */
+public class TrackedTwo {
+
+	public TrackedTwo(TrackedOne trackedOne) {
+		_trackedOne = trackedOne;
+	}
+
+	public TrackedOne getTrackedOne() {
+		return _trackedOne;
+	}
+
+	private final TrackedOne _trackedOne;
+
+}
